Implemented a cooldown usage callback, which triggers at the moment the cooldown is used
With this change, the user sees the cooldown usage in the moment it is used.
Aura duration isn't available at this point.
At a point in the future (100ms to 2000ms), the lib receives the data from the player as usual.

Created a comm scheduler, throttling comm messages to one per second.

Modified event timers; instead of cancelling the timer and creating a new with an interval of 2 to 8 seconds, it will now create a timer with an interval of approximately 20 seconds that cannot be cancelled.